Add Entrypoint lock

+++ rootfs/opt/docker/bin/entrypoint.d/20-user-application.sh
@@ -7,21 +7,23 @@ APPLICATION_USER=${APPLICATION_USER:-application}
 APPLICATION_GROUP=${APPLICATION_USER:-application}
 PASSWORD=${PASSWORD:-Admin123!}
 
-# Add group
-groupadd -g "$APPLICATION_GID" "$APPLICATION_GROUP"
+if ! -f /opt/docker/entrypoint.lock; then
+  # Add group
+  groupadd -g "$APPLICATION_GID" "$APPLICATION_GROUP"
 
-# Add user
-useradd -u "$APPLICATION_UID" --home "/home/$APPLICATION_USER" --create-home --shell /bin/bash --no-user-group "$APPLICATION_USER"
+  # Add user
+  useradd -u "$APPLICATION_UID" --home "/home/$APPLICATION_USER" --create-home --shell /bin/bash --no-user-group "$APPLICATION_USER"
 
-# Assign user to group
-usermod -g "$APPLICATION_GROUP" "$APPLICATION_USER"
-usermod -aG sudo "$APPLICATION_USER"
+  # Assign user to group
+  usermod -g "$APPLICATION_GROUP" "$APPLICATION_USER"
+  usermod -aG sudo "$APPLICATION_USER"
 
-# Set password
-echo "$APPLICATION_USER":"$PASSWORD" | chpasswd
+  # Set password
+  echo "$APPLICATION_USER":"$PASSWORD" | chpasswd
 
-# Copy files
-cp /root/.shell-methods.sh /home/$APPLICATION_USER/
-cp /root/.bashrc /home/$APPLICATION_USER/
+  # Copy files
+  cp /root/.shell-methods.sh /home/$APPLICATION_USER/
+  cp /root/.bashrc /home/$APPLICATION_USER/
 
-chown -R "$APPLICATION_USER":"$APPLICATION_GROUP" /home/$APPLICATION_USER
+  chown -R "$APPLICATION_USER":"$APPLICATION_GROUP" /home/$APPLICATION_USER
+fi
