+++ rootfs/opt/docker/bin/entrypoint.d/30-xfce4.sh
@@ -3,7 +3,9 @@
 # Fix special user permissions
 APPLICATION_USER=${APPLICATION_USER:-application}
 
-echo 'XFCE_PANEL_MIGRATE_DEFAULT=1' > /etc/environment
+if ! -f /opt/docker/entrypoint.lock; then
+  echo 'XFCE_PANEL_MIGRATE_DEFAULT=1' > /etc/environment
 
-sudo rsync -a --relative /root/./.config/xfce4 /home/${APPLICATION_USER}/
-sudo chown -R ${APPLICATION_USER}:${APPLICATION_USER} /home/${APPLICATION_USER}
+  sudo rsync -a --relative /root/./.config/xfce4 /home/${APPLICATION_USER}/
+  sudo chown -R ${APPLICATION_USER}:${APPLICATION_USER} /home/${APPLICATION_USER}
+fi
